Tests/LibWeb/Text/expected/selection-drag-across-image-and-heading.txt

selected heading text: true
selected image: true
